src: build without compiler warnings
  * src/fcnn.cpp: count the loop indices in size_t.
  * src/editDistance.cc (editDistance): count the distance vector in
    size_t.
  * src/fcnnpredict.cc (fcnnpredict): initialize output_size.
  * src/gamboost.cpp, src/tree.cpp: mark the helpers that a given including
    oct-file does not call, and say at the top of each file that the unit is
    included by several.
  * src/svmtrain.cc (parse_command_line): terminate the option buffer when
    the option string fills it.  Mark model_file_name.  (print_null,
    print_null_): drop the unused parameter name.
  * src/svmpredict.cc (svmpredict): terminate the option buffer when the
    option string fills it.  Set error_msg, which the model reader never
    writes.  (predict): mark the fall through.  (print_null): drop the
    unused parameter name.
  * src/svm_model_octave.cc (octave_matrix_to_model): remove num_of_fields
    and sc, and mark msg.  Mark field_names.
  * src/libsvmwrite.cc (write): remove three unused declarations.

+++ src/editDistance.cc
@@ -624,7 +624,7 @@ insertions, deletions, and substitutions required to convert document \
         }
       }
       // Preallocate the distance vector and expand as necessary
-      int sz = szA;
+      size_t sz = szA;
       if (szA == 1 && szB != 1)
       {
         sz = szB;
